Fix sqlite error type and HTTP response code for duplicate username
The official implementation uses 402 response code with a specific error
message when a username is already registered. This change ensures that
the KOReader app shows the correct message when a duplicate username is
registered.

aiosqlite returns an IntegrityError when a violation of the UNIQUE
condition is detected. Catch that and return the error appropriately.